pyproject.toml
[project]
name = "omsg"
version = "0.1.0"
description = "A phone-first messenger backend foundation."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "sqlalchemy>=2.0.35",
  "alembic>=1.18.0",
  "pydantic-settings>=2.6.0",
  "PyJWT>=2.9.0",
  "psycopg[binary]>=3.2.0",
  "python-multipart>=0.0.12",
  "redis>=5.0.0",
  "pytest>=8.3.0",
  "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.api", "app.api.routers", "app.core", "app.models", "app.schemas", "app.services", "app.realtime"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
